Use branch of jekyll-sass-converter to fix deprecation warning
This fixes the deprecation warning (see below) which appears on running
a fresh `bundle install`. This branch contains the changes described in
this PR [1]. The only outstanding problems seem to be Windows-related,
so I think it's safe to use it.

We should be able to switch back to the release version (required by
default by the jekyll gem) when this PR is merged and released.

    Ruby Sass is deprecated and will be unmaintained as of 26 March 2019.

    * If you use Sass as a command-line tool, we recommend using Dart Sass, the new
      primary implementation: https://sass-lang.com/install

    * If you use Sass as a plug-in for a Ruby web framework, we recommend using the
      sassc gem: https://github.com/sass/sassc-ruby#readme

    * For more details, please refer to the Sass blog:
      http://sass.logdown.com/posts/7081811

[1]: https://github.com/jekyll/jekyll-sass-converter/pull/75