Link to forum instead of mailing list
We don't have a process to ensure that emails to our internal list
(contact@coops.tech) are replied to in a timely manner/at all. At the
moment we appear to have 5 legitimate emails from the 13 March that
don't appear to have replies.

I hope that encouraging people to post to the forum will increase the
chance that they receive a reply.

I think there are further improvements we can make, by clarifying the
joining process for example, but this is a good enough start.