Remove duplicate Kaspersky client
Although no coops seem to be referencing this client, there were already
lowercase versions of the relevant files, i.e. `_clients/kaspersky.md`
and `images/clients/kaspersky.svg`. So the uppercase version,
`Kaspersky.md` was effectively a duplicate.

This led to the following warning messages on case-insensitive
filesystems, making it hard to work on the project on such systems:

> warning: the following paths have collided (e.g. case-sensitive paths
> on a case-insensitive filesystem) and only one from the same colliding
> group is in the working tree

In this commit, I've deleted `Kaspersky.md` which didn't contain
anything different to `kaspersky.md`. I think this is the right thing to
do, because ideally all the markdown files should be lowercase in case
they are used in URL paths.

There are other similar issues, but I plan to address them in separate
commits.