Remove duplicate Power to Change client
Both Blake House and Open Data Services were already using lowercase
version of the Power to Change client markdown and image files, i.e.
`_clients/power-to-change.md` and `images/clients/power-to-change.png`.
So the uppercase versions of the files, i.e.
`_clients/Power-to-Change.md` and `images/clients/Power-to-Change.png`,
were effectively duplicates. 

This led to the following warning messages on case-insensitive
filesystems, making it hard to work on the project on such systems:

> warning: the following paths have collided (e.g. case-sensitive paths
> on a case-insensitive filesystem) and only one from the same colliding
> group is in the working tree

In this commit, I've added the website from `Power-to-Change.md` into
`power-to-change.md`, deleted `Power-to-Change.md` &
`Power-to-Change.png`, and changed the reference in `dot-project.md`
from `Power-to-Change` to `power-to-change`. I think this is the right
thing to do, because ideally all the markdown files should be lowercase
in case they are used in URL paths.