Upgrade ruby and gems
Ruby 2.6.6 has reached the end-of-life and I had some trouble
installing the dependencies locally as a result. I've updated to the
latest version of ruby and regenerated the Gemfile.lock with the
latest versions of the two gems we depend on, and their dependencies.

Everything looks to still be working in terms of site generation.