Strip out CSS, JS and extra markup
The site had around 7000 lines of CSS and 23,000 lines of JS. It also
used a framework called "Zurb" to build the layout. We're planning to
rebrand the website and to make that work easier I think it would help
me (and hopefully those that come after me) to strip out all of these
technologies and recreate the site with a simple stylesheet.

By doing so I think we lose a couple of features, such as the zooming hover
states on the grids of images. Some features, like the "sticky" header
are probably better implemented nowadays in pure CSS.

Given that the site is maintained by volunteers I think the simpler we
can make it the better.