Improve look of co-op logos
Before this commit the logos were a fairly garish mixture of different
colours, and many of the logos didn't work very well on a white
background.

In keeping with the new brand guidelines we've applied a greyscale
filter to the logos, and Natasha has created some variants of the
existing logos that work better on a black background. I've decided to
keep all the original logos and their variants in case we want to
switch again later, and chosen which one to used based on my taste.

This seems to improve things and we can ask co-ops to provide us with
higher res logos that work on black backgrounds later if we want to
improve things further.

Co-authored-by: Natasha Natarajan <natasha@outlandish.com>