Strip out unwanted CSS
I think it's probably easier to start from a cleaner slate for this
redesign, rather than try to incrementally refactor towards the new
design. Some of the base styles are fine and I've retained some of the
CSS variables. I've also kept the header as it probably will be useful
to refactor rather than completely re-do this.