Style the co-op pages to match the designs
I've decided to style these pages to look like "modals" for now, but I
haven't implemented them as such yet. I think we might be able to add
some progressive enhancement to retain the scroll position when a
co-op logo is clicked to make it feel more like a separate modal
dialog is opening.