Remove unused images
There appear to be no references to these images in the current output
in `_site` after running `jekyll build`. Over time as we've worked on
a refactored the site, adding and removing co-ops and so on I think
these images have become orphaned.

Removing them makes everything a bit easier to understand.

I've visted each page in turn in the browser with developer tools
enabled and there doesn't appear to be any 404 status codes for any
requests on any page.